requirements.txt
absl-py==1.4.0
albumentations==1.3.1
asgiref==3.7.2
astunparse==1.6.3
attrs==23.1.0
beautifulsoup4==4.12.2
bidict==0.22.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cmake==3.27.2
coloredlogs==15.0.1
contourpy==1.1.0
cycler==0.11.0
Cython==3.0.0
deepface==0.0.79
Django==4.2.4
django-filter==23.2
djangorestframework==3.14.0
dlib==19.24.2
easydict==1.10
exceptiongroup==1.1.3
face-recognition==1.3.0
face-recognition-models==0.3.0
faiss-cpu==1.7.4
filelock==3.12.2
fire==0.5.0
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SocketIO==5.3.5
flatbuffers==23.5.26
fonttools==4.42.1
gast==0.4.0
gdown==4.7.1
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.57.0
gunicorn==21.2.0
h11==0.14.0
h5py==3.9.0
humanfriendly==10.0
idna==3.4
imageio==2.31.1
imutils==0.5.4
insightface==0.7.3
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
keras==2.13.1
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
lit==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
mtcnn==0.1.1
mutagen==1.46.0
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
onnx==1.14.1
onnxruntime==1.15.1
onnxruntime-gpu==1.15.1
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.76
opt-einsum==3.3.0
outcome==1.2.0
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
prettytable==3.8.0
protobuf==4.24.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pygame==2.5.1
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-engineio==4.6.1
python-socketio==5.8.0
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
qudida==0.0.4
requests==2.31.0
requests-oauthlib==1.3.1
retina-face==0.0.13
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.2
selenium==4.11.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.33.0
termcolor==2.3.0
threadpoolctl==3.2.0
tifffile==2023.8.12
torch==2.0.1
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.10.3
triton==2.0.0
ttkthemes==3.2.2
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
wcwidth==0.2.6
Werkzeug==2.3.7
wrapt==1.15.0
wsproto==1.2.0
